+++ src/main/java/org/homemade/elo/entities/dto/OpponentPlayerDetails.java
@@ -1,7 +1,6 @@
 package org.homemade.elo.entities.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -11,11 +10,11 @@ import lombok.NoArgsConstructor;
 public class OpponentPlayerDetails {
 	private static final String MESSAGE = "  %s";
 	private static final String EXTENDED_MESSAGE = "  %s (%d times)";
-	private int id;
+	private long id;
 	private String name;
 	private int matches;
 
-	public OpponentPlayerDetails(int id, String name) {
+	public OpponentPlayerDetails(long id, String name) {
 		this.id = id;
 		this.name = name;
 		this.matches = 1;
