+++ src/main/java/org/homemade/elo/entities/dto/PlayerDetails.java
@@ -1,12 +1,11 @@
 package org.homemade.elo.entities.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.util.HashMap;
 import java.util.Map;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Getter
@@ -17,14 +16,14 @@ public class PlayerDetails {
 	private static final String PLAYER_WON_MATCHES = "* Won matches against:";
 	private static final String PLAYER_LOST_MATCHES = "* Was beaten by:";
 	private String name;
-	private Map<Integer, OpponentPlayerDetails> beat = new HashMap<>();
-	private Map<Integer, OpponentPlayerDetails> lostFrom = new HashMap<>();
+	private Map<Long, OpponentPlayerDetails> beat = new HashMap<>();
+	private Map<Long, OpponentPlayerDetails> lostFrom = new HashMap<>();
 
 	public PlayerDetails(String name) {
 		this.name = name;
 	}
 
-	public void addBeatByCurrentPlayer(int id, String name) {
+	public void addBeatByCurrentPlayer(long id, String name) {
 		if (this.beat.containsKey(id)) {
 			this.beat.get(id).incMatches();
 		} else {
@@ -32,7 +31,7 @@ public class PlayerDetails {
 		}
 	}
 
-	public void addCurrentLostFromPlayer(int id, String name) {
+	public void addCurrentLostFromPlayer(long id, String name) {
 		if (this.lostFrom.containsKey(id)) {
 			this.lostFrom.get(id).incMatches();
 		} else {
